Fix configuration circular dependencies
